backend/requirements.txt
django>=4.2
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
psycopg2-binary>=2.9
python-dotenv>=1.0
PyPDF2>=3.0
python-docx>=1.1
Pillow>=10.0
reportlab>=4.0
